fix #725: View.AddAnimatorEndListener 추가 및 적용
